wifi: rtw88: Shared module for rtw8723x devices
The already supported 8723d chip is very similar to 8703b/8723cs,
split code that can be shared into a new module. The spec definition
tables are combined into a struct so we only need one EXPORT_SYMBOL
for them all.

Acked-by: Ping-Ke Shih <pkshih@realtek.com>
Tested-by: Pavel Machek <pavel@ucw.cz>
Link: https://msgid.link/20240311103735.615541-2-fiona.klute@gmx.de
(cherry picked from commit ff88b74882669e8b1931730b2401dbc2bece1355)
